Stop the bot logger from treating finished text as a format string
PlayerbotAIConfig::log(fileName, str, ...) handed str straight to vfprintf. Of
its 68 call sites, 67 pass a line they have already assembled - the usual shape
being log("bot_events.csv", out.str().c_str()). A percent sign anywhere in that
line, from a bot name, a mob name or an item name, was read as a conversion
specifier and vfprintf reached for an argument nobody passed.

glibc prints nonsense and carries on, which is why this never showed on Linux.
The Microsoft runtime calls the invalid-parameter handler instead, so the
Windows server died with "invalid-parameter (0xc0000420)" followed by SIGABRT as
soon as the first bots came up and AttackAnythingAction logged its first event.
The minidump named PlayerbotAIConfig::log -> _vfprintf_l ->
MangosdInvalidParameterHandler outright.

The signature was the trap, so it is split in two: log() now writes the line
verbatim through fputs, and logf() keeps the varargs form for the single caller
that genuinely formats. On GCC and Clang logf carries the format attribute, so
the next mistake of this kind is a compiler warning rather than a crash dump.

+++ src/modules/PlayerBots/playerbot/TravelNode.cpp
@@ -3247,7 +3247,7 @@ void TravelNodeMap::printNodeStore()
     sPlayerbotAIConfig.log(nodeStore, "    public:");
     sPlayerbotAIConfig.log(nodeStore, "    static void loadNodes()");
     sPlayerbotAIConfig.log(nodeStore, "    {");
-    sPlayerbotAIConfig.log(nodeStore, "        TravelNode** nodes = new TravelNode*[%d];", anodes.size());
+    sPlayerbotAIConfig.logf(nodeStore, "        TravelNode** nodes = new TravelNode*[%d];", (int)anodes.size());
 
     for (uint32 i = 0; i < anodes.size(); i++)
     {
